fix: src files to TS
