--- src/App.js
@@ -1,37 +1,0 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import { ToastContainer } from 'react-toastify'
-import 'react-toastify/dist/ReactToastify.css'
-import Navbar from './components/Navbar'
-import PrivateRoute from './components/PrivateRoute'
-import Explore from './pages/Explore'
-import Chat from './pages/Chat/Chat'
-import Home from './pages/Home'
-import Profile from './pages/Profile'
-import SignIn from './pages/SignIn'
-import SignUp from './pages/SignUp'
-import ForgotPassword from './pages/ForgotPassword'
-function App() {
-  
-  return (
-    <>
-      <Router>
-        <Routes>
-        <Route path='/' element={<Home />} />
-          <Route path='/explore' element={<Explore />} />
-          <Route path='/chat' element={<Chat />} />
-          <Route path='/profile' element={<PrivateRoute />}>
-            <Route path='/profile' element={<Profile />} />
-          </Route>
-          <Route path='/sign-in' element={<SignIn />} />
-          <Route path='/sign-up' element={<SignUp />} />
-          <Route path='/forgot-password' element={<ForgotPassword />} />
-        </Routes>
-        <Navbar />
-      </Router>
-
-      <ToastContainer />
-    </>
-  )
-}
-
-export default App
